remove_article_first_line/requirements.txt
contentful-management>=2.11.0
python-dotenv>=0.19.0

